refactor: 공통되는 기능을 Common 모듈로 분리
