+++ common/src/main/kotlin/io/livestreaming/api/common/domain/Image.kt
@@ -1,4 +1,4 @@
-package io.livestreaming.api.member.domain
+package io.livestreaming.api.common.domain
 
 class Image private constructor(
     val url: URL,
