+++ common/src/main/kotlin/io/livestreaming/api/common/domain/MemberId.kt
@@ -1,4 +1,4 @@
-package io.livestreaming.api.member.domain
+package io.livestreaming.api.common.domain
 
 class MemberId private constructor(val value: String) {
