+++ common/src/main/kotlin/io/livestreaming/api/common/domain/URL.kt
@@ -1,4 +1,4 @@
-package io.livestreaming.api.member.domain
+package io.livestreaming.api.common.domain
 
 class URL private constructor(val value: String) {
