+++ common/src/test/kotlin/io/livestreaming/api/common/domain/URLTest.kt
@@ -1,8 +1,7 @@
-package io.livestreaming.api.member.domain
+package io.livestreaming.api.common.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
-import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
