+++ member/build.gradle.kts
@@ -5,6 +5,7 @@ allOpen {
 }
 
 dependencies {
+    implementation(project(":common"))
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
 }
