+++ member/src/main/kotlin/io/livestreaming/api/member/MemberApplication.kt
@@ -3,7 +3,7 @@ package io.livestreaming.api.member
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = ["io.livestreaming.api.common"])
 class UserApplication
 
 fun main(args: Array<String>) {
