+++ member/src/main/kotlin/io/livestreaming/api/member/application/UpdateMemberPasswordCommand.kt
@@ -1,6 +1,6 @@
 package io.livestreaming.api.member.application
 
-import io.livestreaming.api.member.domain.MemberId
+import io.livestreaming.api.common.domain.MemberId
 import io.livestreaming.api.member.domain.Password
 
 data class UpdateMemberPasswordCommand(
