+++ member/src/main/kotlin/io/livestreaming/api/member/application/UpdateMemberUseCase.kt
@@ -1,6 +1,6 @@
 package io.livestreaming.api.member.application
 
-import io.livestreaming.api.member.domain.MemberId
+import io.livestreaming.api.common.domain.MemberId
 
 interface UpdateMemberUseCase {
     fun updatePassword(command: UpdateMemberPasswordCommand): MemberId
