+++ member/src/main/kotlin/io/livestreaming/api/member/util/id/MemberIdGenerator.kt
@@ -1,6 +1,6 @@
 package io.livestreaming.api.member.util.id
 
-import io.livestreaming.api.member.domain.MemberId
+import io.livestreaming.api.common.domain.MemberId
 import io.livestreaming.api.member.util.uuid.UuidProvider
 import org.springframework.stereotype.Component
